pyproject.toml
[project]
name = "education-story-generator"
version = "1.0.0"
description = "AI-powered educational story generator and content API for lesson plans"
authors = [
    {name = "Education Story Creator", email = "dev@education-stories.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "mistralai>=1.5.2",
    "requests>=2.32.4",
    "pytesseract>=0.3.13",
    "trustcall>=0.0.26",
    "langchain-openai>=0.3.32",
    "pymupdf>=1.26.4",
    "pymongo>=4.14.1",
    "pydantic[email]>=2.11.7",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "bcrypt==4.0.1",
]

[project.urls]
Homepage = "https://github.com/education-story-creation"
Repository = "https://github.com/education-story-creation"

[project.scripts]
story-generator = "story_generator:main"
difficult-word-api = "difficult_word:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "story_generator.py",
    "difficult_word.py",
    "README.md",
]

[tool.uv]
dev-dependencies = []
